+++ robospock/src/main/groovy/org/robospock/GradleRoboSpecification.groovy
@@ -1,11 +1,8 @@
-package pl.polidea.robospock
+package org.robospock
 
 import org.junit.runner.RunWith
-import org.robospock.GradleRoboSputnik
 import spock.lang.Specification
 
-
 @RunWith(GradleRoboSputnik)
-@Deprecated
 class GradleRoboSpecification extends Specification {
 }
