+++ robospock/src/main/groovy/org/robospock/RoboSpecification.groovy
@@ -1,9 +1,7 @@
-package pl.polidea.robospock
+package org.robospock
 
 import org.junit.runner.RunWith
-import org.robospock.RoboSputnik
 import spock.lang.Specification
 
 @RunWith(RoboSputnik)
-@Deprecated
 class RoboSpecification extends Specification {}
