+++ robospock/src/main/java/org/robospock/ParallelUniverseCompat.java
@@ -1,4 +1,4 @@
-package pl.polidea.robospock.internal;
+package org.robospock;
 
 import android.app.Application;
 import android.content.Context;
